Calculate yawspeeds in TasTools
Prior to this commit, yawspeed calculations are always done by genlegit.py.  It
accomplishes this by taking the difference between the yaw angles of
consecutive frames.  This can be a problem if the game modifies the yaw angle
(e.g. when the player is standing on some rotating object).  This commit fixes
this problem.