Added cl_lgagst_duck cvar
If the value of this cvar is 1, then lgagst will calculate the air/ground
accels using the modified value of sv_maxspeed (scaled down by 0.333) when the
player is ducked.  Otherwise, it will always use the original sv_maxspeed.

The AGST is different depending on the value of this cvar.